-add admin features

+++ server/src/routes/api/v1/userPreferencesRouter.js
@@ -10,28 +10,48 @@ const userPreferencesRouter = new express.Router({ mergeParams: true });
 
 userPreferencesRouter.post("/", async (req, res) => {
     const { name } = req.body;
-    try {
-        const user = await User.query().findOne({ id: req.user.id });
-        await AddPreferenceToDatabase.addOnePreference(name, user);
-        const preferences = await user.$relatedQuery("preferences");
-        const serializedPreferences = PreferenceSerializer.getSummaryOfArray(preferences);
-        return res.status(201).json({ preferences: serializedPreferences });
-    } catch (error) {
-        return res.status(422).json({ errors: error });
+    if (req.user.username === "admin") {
+        try {
+            await AddPreferenceToDatabase.adminAddPreference(name);
+            const preferences = await Preference.query();
+            const serializedPreferences = PreferenceSerializer.getSummaryOfArray(preferences);
+            return res.status(201).json({ preferences: serializedPreferences });
+        } catch (error) {
+            return res.status(422).json({ errors: error });
+        }
+    } else {
+        try {
+            const user = await User.query().findOne({ id: req.user.id });
+            await AddPreferenceToDatabase.addOnePreference(name, user);
+            const preferences = await user.$relatedQuery("preferences");
+            const serializedPreferences = PreferenceSerializer.getSummaryOfArray(preferences);
+            return res.status(201).json({ preferences: serializedPreferences });
+        } catch (error) {
+            return res.status(422).json({ errors: error });
+        }
     }
 });
 
 userPreferencesRouter.delete("/:preferenceId", async (req, res) => {
-    try {
-        const preferenceToDelete = await Preference.query().findById(req.params.preferenceId);
-        const userPreferenceToDelete = await UserPreference.query().findOne({
-            preferenceId: req.params.preferenceId,
-            userId: req.user.id,
-        });
-        await userPreferenceToDelete.$query().delete();
-        return res.status(200).json({ preference: preferenceToDelete });
-    } catch (err) {
-        return res.status(500).json({ errors: err });
+    if (req.user.username === "admin") {
+        try {
+            const preferenceToDelete = await Preference.query().findById(req.params.preferenceId);
+            return res.status(200).json({ preference: preferenceToDelete });
+        } catch (err) {
+            return res.status(500).json({ errors: err });
+        }
+    } else {
+        try {
+            const preferenceToDelete = await Preference.query().findById(req.params.preferenceId);
+            const userPreferenceToDelete = await UserPreference.query().findOne({
+                preferenceId: req.params.preferenceId,
+                userId: req.user.id,
+            });
+            await userPreferenceToDelete.$query().delete();
+            return res.status(200).json({ preference: preferenceToDelete });
+        } catch (err) {
+            return res.status(500).json({ errors: err });
+        }
     }
 });
